Improve the project organization
