+++ .env.example
@@ -1,0 +1,4 @@
+DISCORD_TOKEN=TOKEN_HERE
+
+FOLDER_SERIES=path_here
+FOLDER_MOVIES=path_here
